requirements.txt
flask==2.0.1
requests==2.26.0
langdetect==1.0.9
iso639==0.1.4
python-dotenv==0.19.0
flask-socketio==5.1.1
werkzeug==2.0.1 